Remove APIClient as it looks like it's not in use
Reasons I think is not in use (beyond instantiation):

1. In Shopicy/cli/packages/ui-extensions-server-kit ExtensionServerClient is the only code to instantiate APIClient.  BUT! it does nothing with the instantiated code (e.g: use one of APIClient 's methods).
2. In Shopicy/cli/packages/ui-extensions-dev-console nothing uses the APIClient directly.
3. In Shopify/web nothing seemed to use the APIClient directly.

As part of removing APIClient We get to:

1. Move filterExtensionsBySurface directly into ExtensionsServerClient, which is less overhead than a utilities folder and an index file (which suggests it's reused when it's not)
2. Simplify the typing of filterExtensionsBySurface.  Previously this had to satisfy two slightly different types.  Now it only has to simplify one.
