+++ packages/ui-extensions-server-kit/src/ExtensionServerClient/ExtensionServerClient.ts
@@ -1,8 +1,7 @@
 /* eslint-disable no-console */
-import {filterExtensionsBySurface} from './utilties'
-import {APIClient} from './APIClient'
+import {Surface} from './types.js'
 import {isValidSurface} from '../utilities'
-import {DeepPartial} from '../types'
+import {DeepPartial, ExtensionPayload} from '../types'
 
 export class ExtensionServerClient implements ExtensionServer.Client {
   public id: string
@@ -82,17 +81,6 @@ export class ExtensionServerClient implements ExtensionServer.Client {
     this.connection?.send(JSON.stringify({event: 'dispatch', data: {type: event, payload: data}}))
   }
 
-  protected initializeApiClient() {
-    let url = ''
-    if (this.options.connection.url) {
-      // eslint-disable-next-line node/no-unsupported-features/node-builtins
-      const socketUrl = new URL(this.options.connection.url)
-      socketUrl.protocol = socketUrl.protocol === 'ws:' ? 'http:' : 'https:'
-      url = socketUrl.origin
-    }
-    this.api = new APIClient(url, this.options.surface)
-  }
-
   protected initializeConnection() {
     if (!this.connection) {
       return
@@ -138,10 +126,6 @@ export class ExtensionServerClient implements ExtensionServer.Client {
       return
     }
 
-    if (!this.api || this.api.url !== this.connection.url) {
-      this.initializeApiClient()
-    }
-
     if (!connectWebsocket) {
       return
     }
@@ -176,4 +160,31 @@ function getValidatedOptions<TOptions extends DeepPartial<ExtensionServer.Option
     delete options.surface
   }
   return options
+}
+
+function filterExtensionsBySurface(extensions: ExtensionPayload[], surface: Surface | undefined): ExtensionPayload[] {
+  if (!surface) {
+    return extensions
+  }
+
+  return extensions.filter((extension) => {
+    if (extension.surface === surface) {
+      return true
+    }
+
+    if (Array.isArray(extension.extensionPoints)) {
+      const extensionPoints: (string | {surface: Surface; [key: string]: any})[] = extension.extensionPoints
+      const extensionPointMatchingSurface = extensionPoints.filter((extensionPoint) => {
+        if (typeof extensionPoint === 'string') {
+          return false
+        }
+
+        return extensionPoint.surface === surface
+      })
+
+      return extensionPointMatchingSurface.length > 0
+    }
+
+    return false
+  })
 }
