--- packages/ui-extensions-server-kit/src/ExtensionServerClient/utilties/index.ts
@@ -1,1 +1,0 @@
-export * from './filterExtensionsBySurface'
